requirements.txt
torch>=2.1.0
transformers>=4.40.0
python-dotenv>=0.21.0
bitsandbytes
accelerate>=0.27.0
sentencepiece
openpyxl
langchain 
langchain-huggingface
langraph
sqlglot
psycopg2-binary
scipy
numpy
matplotlib
pandas